drivers/docker: fix a hostConfigMemorySwappiness panic (#18238)
cgroupslib.MaybeDisableMemorySwappiness returned an incorrect type, and was
incorrectly typecast to int64 causing a panic on non-linux and non-windows hosts.

+++ client/lib/cgroupslib/default.go
@@ -11,6 +11,6 @@ func LinuxResourcesPath(string, string) string {
 }
 
 // MaybeDisableMemorySwappiness does nothing on non-Linux systems
-func MaybeDisableMemorySwappiness() *int {
+func MaybeDisableMemorySwappiness() *uint64 {
 	return nil
 }
